Assignment 1 almost complete. Need links for a list of all rounds for all decks they've completed. See release 0 directions

+++ flash_cards/db/migrate/20140425114246_create_rounds.rb
@@ -3,7 +3,7 @@ class CreateRounds < ActiveRecord::Migration
     create_table :rounds do |t|
       t.belongs_to :deck
       t.belongs_to :user
-      t.integer  :score, default: 0
+      t.float  :score, default: 0
 
       t.timestamps
     end
